SQLitePersistenceService Done Fix Tests calculate

+++ TUI_SimulatorFlightCosts/TUI_SimulatorFlightCosts_AcceptanceTests/Calculate_Flight_Distance_And_Fuel_Needed_Steps.cs
@@ -55,13 +55,13 @@ namespace TUI_SimulatorFlightCosts_AcceptanceTests
 
 
         [Then(@"the travel distance was (.*) km")]
-        public void ThenTheTravelDistanceWasKm(int distanceExpected)
+        public void ThenTheTravelDistanceWasKm(double distanceExpected)
         {
             _distanceKm.Should().Be(distanceExpected);
         }
         
         [Then(@"the total fuel consumption was (.*) L")]
-        public void ThenTheTotalFuelConsumptionWasL(int fuelConsumptionExpected)
+        public void ThenTheTotalFuelConsumptionWasL(double fuelConsumptionExpected)
         {
             _fuelConsumption.Should().Be(fuelConsumptionExpected);
         }
